issues/18: reorganizing decorators
